requirements.txt
# Flask core
Flask==1.1.4
Werkzeug==1.0.1
Jinja2==2.11.3
# Jinja2 2.11.x requires soft_unicode (removed in MarkupSafe 2.1+)
MarkupSafe==2.0.1
itsdangerous==1.1.0
click==7.1.2

# Database
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
psycopg2-binary==2.9.9

# REST API
flask-restful==0.3.9
aniso8601==9.0.1
six==1.16.0

# Serialization / Validation
flask-marshmallow==0.14.0
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1

# Auth
Flask-JWT-Extended==3.25.1
PyJWT==1.7.1

# Env vars
python-dotenv==0.21.0

# Production server
gunicorn==21.2.0
